moved get_values to admin, removed extra LazyConfig
